Rename deprecated select export
