+++ src/index.ts
@@ -47,7 +47,6 @@ export { default as ColorPicker, COLORS } from './components/color-picker'
 export { default as KeyboardShortcut } from './components/keyboard-shortcut'
 export { default as KeyCapturer, SUPPORTED_KEYS } from './components/key-capturer'
 export { default as ProgressBar } from './components/progress-bar'
-export { default as Select } from './components/select'
 export { default as Time } from './components/time'
 export { Notification } from './components/notification/notification'
 export { Tooltip } from './components/tooltip'
@@ -61,3 +60,4 @@ export * from './components/menu'
 export { default as DeprecatedButton } from './components/deprecated-button'
 export { default as DeprecatedDropdown } from './components/deprecated-dropdown'
 export { default as DeprecatedInput } from './components/deprecated-input'
+export { default as DeprecatedSelect } from './components/deprecated-select'
